[FIX] 개념 생성 api 오류 해결 Final Final

+++ src/main/java/com/sciencematch/sciencematch/service/AdminService.java
@@ -89,7 +89,7 @@ public class AdminService {
     public Long postConcept(ConceptPostDto conceptPostDto) throws IOException {
         Chapter chapter = chapterRepository.getChapterById(conceptPostDto.getChapterId());
         Concept postConcept = conceptRepository.getByChapterId(conceptPostDto.getChapterId());
-        if (postConcept.getImage() != null) {
+        if (postConcept != null && postConcept.getImage() != null) {
             s3Service.deleteFile(postConcept.getImage());
             conceptRepository.delete(postConcept);
         }
